+++ src/pages/Auth/Login.jsx
@@ -1,114 +1,11 @@
-import { useState } from "react";
-import { Link } from "react-router-dom";
-import useLoadingStore from "../../store/loadingstore";
-import { useNavigate } from "react-router";
-import { Eye, EyeOff } from "lucide-react"; // Add this import
+import Login from '../../components/Auth/Login'
 
-const Login = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [passwordError, setPasswordError] = useState(false);
-  const [showPassword, setShowPassword] = useState(false); // Add this state
-  const { start, stop } = useLoadingStore();
-  const navigate = useNavigate();
-
-  const handleEmailChange = (e) => {
-    setEmail(e.target.value);
-  };
-
-  const handlePasswordChange = (e) => {
-    setPassword(e.target.value);
-    setPasswordError(!/^(?=.*[A-Z]).*$/.test(e.target.value));
-  };
-
-  const togglePasswordVisibility = () => {
-    setShowPassword(!showPassword);
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    start();
-    // Handle form submission here
-
-    setTimeout(() => {
-      navigate("/dashboard");
-      stop();
-    }, 2000);
-  };
-
+function LoginPage() {
   return (
-    <div className="min-h-screen bg-gray-100 flex items-center justify-center">
-      <div className="bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4">
-        <h2 className="text-2xl font-bold mb-4">Welcome Back!</h2>
-        <form onSubmit={handleSubmit}>
-          <div className="mb-4">
-            <label
-              htmlFor="email"
-              className="block text-gray-700 text-sm font-bold mb-2"
-            >
-              Email Address
-            </label>
-            <input
-              type="email"
-              id="email"
-              placeholder="Enter email here"
-              className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-              value={email}
-              onChange={handleEmailChange}
-            />
-          </div>
-          <div className="mb-6">
-            <label
-              htmlFor="password"
-              className="block text-gray-700 text-sm font-bold mb-2"
-            >
-              Password
-            </label>
-            <div className="relative">
-              <input
-                type={showPassword ? "text" : "password"}
-                id="password"
-                placeholder="Enter Password here"
-                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline pr-10"
-                value={password}
-                onChange={handlePasswordChange}
-              />
-              <button
-                type="button"
-                className="absolute inset-y-0 right-0 pr-3 flex items-center"
-                onClick={togglePasswordVisibility}
-              >
-                {showPassword ? (
-                  <EyeOff className="h-5 w-5 text-gray-500" />
-                ) : (
-                  <Eye className="h-5 w-5 text-gray-500" />
-                )}
-              </button>
-            </div>
-            {passwordError && password && (
-              <p className="text-red-500 text-xs mt-1">
-                Password must contain at least one uppercase letter.
-              </p>
-            )}
-          </div>
-          <div className="flex items-center justify-between">
-            <button
-              type="submit"
-              className="bg-blue-500 hover:bg-blue-700 w-full text-white font-bold py-2 px-4 rounded"
-            >
-              Log In
-            </button>
-          </div>
-        </form>
-        <p className="text-center text-gray-600 text-xs mt-4">
-          Don&apos;t have an account?{" "}
-          <Link to="/signup" className="text-blue-500">
-            Sign Up
-          </Link>
-        </p>
-      </div>
+    <div>
+      <Login />
     </div>
-  );
-};
+  )
+}
 
-export default Login;
+export default LoginPage
